requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
yfinance>=0.2.40
pandas>=2.2.0
scipy>=1.13.0
openai>=1.50.0
